refactor: create generate API doc separated file
