+++ src/Server/Application/Shared/Middlewares/HealthCheck/HealthCheck.ts
@@ -1,8 +1,10 @@
 import { Response } from 'express'
 import { Controller, Get, Res } from 'routing-controllers'
+import { Service } from 'typedi'
 import { StatusCode } from '../../APIs/Enums/Status'
 
 @Controller('/health')
+@Service()
 export default class HealthCheck {
   @Get()
   ExecuteAsync(@Res() res : Response) {
